fix: init new file
