+++ core/src/ten_manager/designer_frontend/src/components/app-bar/menu/app-menu.tsx
@@ -94,6 +94,11 @@ export function AppMenu(props: {
       metadata: {
         type: EDefaultWidgetType.AppsManager,
       },
+      popup: {
+        width: 0.5,
+        height: 0.8,
+        maxWidth: 800,
+      },
     });
   };
