+++ core/src/ten_manager/designer_frontend/src/flow/node/graph/context-menu.tsx
@@ -97,6 +97,11 @@ export const ContextMenuItems = (props: { graph: IGraph }) => {
       metadata: {
         type: EDefaultWidgetType.AppsManager,
       },
+      popup: {
+        width: 0.5,
+        height: 0.8,
+        maxWidth: 800,
+      },
     });
   };
